pyproject.toml
[project]
name = "cv-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.8.0",
    "pillow>=11.1.0",
    "pydantic>=2.11.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
